+++ staff/delete_reservation_data.php
@@ -5,24 +5,26 @@ if(isset($_POST['deleteReservation']))
 {
     $reservationID = $_POST['deleteSearchReservation'];
 
-    $sqlCommand = "DELETE FROM reservation WHERE reservation_ID = '$reservationID'";
+    if($reservationID != ""){
+        $sqlCommand = "DELETE FROM reservation WHERE reservation_ID = '$reservationID'";
 
-    echo $sqlCommand;
+        echo $sqlCommand;
 
 
-    $result = mysqli_query($conn, $sqlCommand);
+        $result = mysqli_query($conn, $sqlCommand);
 
-    if($result==true)
-    {
-        header("Location: staff_homepage.php#viewReservation");
+        if($result==true)
+        {
+            header("Location: staff_homepage.php#viewReservation");
+        }
+        else 
+        {
+            header('Location: staff_homepage.php?ililiem2');
+        }
     }
     else 
     {
-        header('Location: staff_homepage.php?error');
+        header('Location: staff_homepage.php?ililiem2');
     }
-}
-else 
-{
-    header('Location: staff_homepage.php?error');
 }
 ?>
